+++ src/actions/contact.ts
@@ -1,17 +1,20 @@
 "use server"
+import { getContactSchema } from "@/schemas";
 import { ContactType } from "@/schemas/types";
-import { ZodSafeParseResult } from "zod";
+import { getTranslations } from "next-intl/server";
 
-export async function sendMessage(validatedFields: ZodSafeParseResult<ContactType>): Promise<{
-     success?: "message.success",
-     error?: "invalidFields" | "message.sendError"
+export async function sendMessage(values: ContactType): Promise<{
+     success?: string,
+     error?: string
 }>{
+     const t = await getTranslations("validation")
+     const validatedFields = getContactSchema(t).safeParse(values);
      if(!validatedFields.success)
-          return { error: "invalidFields" }
+          return { error: t("invalidFields") }
      const res = await fetch("https://formspree.io/f/myybakel", {
           method: "POST",
           headers: { "Content-Type": "application/json" },
           body: JSON.stringify(validatedFields.data),
      });
-     return res.ok ? { success: "message.success" } : { error: "message.sendError" }
+     return res.ok ? { success: t("message.success") } : { error: t("message.sendError") }
 }